AKS: Intermediate

+++ posts/views.py
@@ -2,6 +2,7 @@ from urllib import quote_plus
 
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.http import Http404
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import render, get_object_or_404, redirect
 
@@ -10,6 +11,8 @@ from .models import Post
 
 
 def post_create(request):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
@@ -53,6 +56,8 @@ def post_list(request):
 
 
 def post_update(request, slug=None):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
@@ -69,6 +74,8 @@ def post_update(request, slug=None):
 
 
 def post_delete(request, slug=None):
+    if not request.user.is_staff or request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Successfully deleted!")
